Go to event page and then back to '/events' page
I need to use XPATH to get the events container because there aren't many unique identifiers on this page. But this XPATH changes; seems there are only two options, so thats what lines 59-64 are for. Then, I loops through events, getting the link to event page for each one. Opening the event page, and then going back to the home page.